- casino: restructure file layout

+++ zoidweb2-casino.sql
@@ -36,8 +36,8 @@ grant all on casino.cardtable to apache;
 
 create table casino.__betlog (
     "id" serial unique not null primary key,
-    "memberid" integer constraint fk_betlog_memberid references engine.__member(id) on update cascade on delete set null,
-    "cardtableid" integer constraint fk_betlog_cardtableid references casino.__cardtable(id) on update cascade on delete set null,
+    "memberid" bigint constraint fk_betlog_memberid references engine.__member(id) on update cascade on delete set null,
+    "cardtableid" bigint constraint fk_betlog_cardtableid references casino.__cardtable(id) on update cascade on delete set null,
     "amount" numeric(10,0) not null,
     "dateposted" timestamptz,
     "description" text
@@ -46,9 +46,9 @@ create table casino.__betlog (
 grant all on casino.__betlog to apache;
 
 create view casino.betlog as
-    select __betlog.*,
-    extract(epoch from dateposted) as datepostedepoch,
-    m1.name as membername
+  select __betlog.*,
+   extract(epoch from dateposted) as datepostedepoch,
+   m1.name as membername
   from casino.__betlog
   LEFT JOIN engine.member m1 ON m1.id = casino.__betlog.memberid
 ;
